PixelSprite Hot Spot support added. You can now set the hot spot (registration point) of any PixelSprite to your own coordinates, or one of 9 pre-defined locations, thus making centering objects a breeze. Scaling the object automatically scales the hot spots.
getDistance updated to add 3rd parameter, fromHotSpot, to control if the distance reported is from the objects x/y (top-left) location or the hot spot.

The PixelSprites default hot spot set to the center of the image. Can be modified with either the hotspot() method, or the hotspotX/hotspotY vars.

Fixed bug in renderer where it would still draw alpha 0 value objects if the visibility was still true.

Added "smoothing" support for scaling.

Updated documentation.